move kind cluster logic to its own file
