+++ pkg/gateway/gateway.go
@@ -1,12 +1,10 @@
 package gateway
 
 import (
-	"bytes"
 	"context"
 	"errors"
 	"fmt"
 	"net"
-	"net/netip"
 	"reflect"
 	"strings"
 	"time"
@@ -21,18 +19,15 @@ import (
 	envoyproxytypes "github.com/envoyproxy/go-control-plane/pkg/cache/types"
 	cachev3 "github.com/envoyproxy/go-control-plane/pkg/cache/v3"
 	resourcev3 "github.com/envoyproxy/go-control-plane/pkg/resource/v3"
-	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
-	"k8s.io/apimachinery/pkg/util/sets"
 	"k8s.io/client-go/tools/cache"
 	"k8s.io/client-go/util/retry"
 	"k8s.io/klog/v2"
 	"k8s.io/utils/ptr"
 
-	"sigs.k8s.io/cloud-provider-kind/pkg/config"
 	"sigs.k8s.io/cloud-provider-kind/pkg/container"
 	gatewayv1 "sigs.k8s.io/gateway-api/apis/v1"
 )
@@ -65,31 +60,9 @@ func (c *Controller) syncGateway(ctx context.Context, key string) error {
 	containerName := gatewayName(c.clusterName, namespace, name)
 	klog.Infof("Syncing Gateway %s, container %s", key, containerName)
 
-	if !container.IsRunning(containerName) {
-		klog.Infof("container %s for gateway is not running", name)
-		if container.Exist(containerName) {
-			if err := container.Delete(containerName); err != nil {
-				return err
-			}
-		}
-	}
-	if !container.Exist(containerName) {
-		klog.V(2).Infof("creating container %s for gateway  %s/%s on cluster %s", containerName, namespace, name, c.clusterName)
-		enableTunnels := c.tunnelManager != nil || config.DefaultConfig.LoadBalancerConnectivity == config.Portmap
-		err := createGateway(c.clusterName, c.xdsLocalAddress, c.xdsLocalPort, gw, enableTunnels)
-		if err != nil {
-			return err
-		}
-
-		// TODO fix this hack
-		time.Sleep(1 * time.Second)
-
-		if err := c.configureContainerNetworking(ctx, containerName); err != nil {
-			if delErr := container.Delete(containerName); delErr != nil {
-				klog.Errorf("failed to delete container %s after networking setup failed: %v", containerName, delErr)
-			}
-			return fmt.Errorf("failed to configure networking for new gateway container %s: %w", containerName, err)
-		}
+	err = c.ensureGatewayContainer(ctx, gw)
+	if err != nil {
+		return fmt.Errorf("failed to ensure gateway container %s: %w", containerName, err)
 	}
 
 	newGw := gw.DeepCopy()
@@ -621,108 +594,4 @@ func createClusterLoadAssignment(clusterName, serviceHost string, servicePort ui
 			},
 		},
 	}
-}
-
-func (c *Controller) getClusterNodeIP(ctx context.Context) (string, error) {
-	nodes, err := c.client.CoreV1().Nodes().List(ctx, metav1.ListOptions{})
-	if err != nil || len(nodes.Items) == 0 {
-		return "", fmt.Errorf("failed to list kubernetes nodes: %w", err)
-	}
-
-	var nodeIP string
-	for _, node := range nodes.Items {
-		if _, ok := node.Labels["node-role.kubernetes.io/control-plane"]; ok {
-			for _, addr := range node.Status.Addresses {
-				if addr.Type == corev1.NodeInternalIP {
-					nodeIP = addr.Address
-					break
-				}
-			}
-		}
-		if nodeIP != "" {
-			break
-		}
-	}
-	if nodeIP == "" {
-		for _, addr := range nodes.Items[0].Status.Addresses {
-			if addr.Type == corev1.NodeInternalIP {
-				nodeIP = addr.Address
-				break
-			}
-		}
-	}
-	if nodeIP == "" {
-		return "", fmt.Errorf("no internal IP found for any node to use as a gateway")
-	}
-	return nodeIP, nil
-}
-
-func (c *Controller) getServiceCIDRs(ctx context.Context) ([]string, error) {
-	serviceCIDRs, err := c.client.NetworkingV1().ServiceCIDRs().List(ctx, metav1.ListOptions{})
-	if err != nil {
-		return nil, fmt.Errorf("failed to list servicecidrs: %w. Is the ServiceCIDR feature gate enabled?", err)
-	}
-	if len(serviceCIDRs.Items) == 0 {
-		return nil, fmt.Errorf("no servicecidrs found in the cluster")
-	}
-
-	cidrs := sets.Set[string]{}
-	for _, serviceCIDRObject := range serviceCIDRs.Items {
-		cidrs.Insert(serviceCIDRObject.Spec.CIDRs...)
-	}
-	if len(cidrs) == 0 {
-		return nil, fmt.Errorf("no CIDRs found in any ServiceCIDR object")
-	}
-	return cidrs.UnsortedList(), nil
-}
-
-func (c *Controller) configureContainerNetworking(ctx context.Context, containerName string) error {
-	var stdout, stderr bytes.Buffer
-
-	installCmd := []string{"sh", "-c", "apt-get update && apt-get install -y iproute2"}
-	klog.Infof("Installing iproute2 in container %s", containerName)
-	if err := container.Exec(containerName, installCmd, nil, &stdout, &stderr); err != nil {
-		klog.Warningf("Failed to install iproute2 in container %s, proceeding anyway: %v", containerName, err)
-	}
-
-	nodeIP, err := c.getClusterNodeIP(ctx)
-	if err != nil {
-		return err
-	}
-	nodeAddr, err := netip.ParseAddr(nodeIP)
-	if err != nil {
-		return err
-	}
-
-	serviceCIDRs, err := c.getServiceCIDRs(ctx)
-	if err != nil {
-		return fmt.Errorf("failed to list servicecidrs: %w", err)
-	}
-	if len(serviceCIDRs) == 0 {
-		return fmt.Errorf("no servicecidrs found in the cluster")
-	}
-
-	var routesAdded int
-	for _, cidr := range serviceCIDRs {
-		prefix, err := netip.ParsePrefix(cidr)
-		if err != nil {
-			return fmt.Errorf("failed to parse CIDR %s: %w", cidr, err)
-		}
-		if prefix.Addr().Is4() != nodeAddr.Is4() {
-			continue
-		}
-
-		cmd := []string{"ip", "route", "replace", cidr, "via", nodeIP}
-		klog.Infof("Adding route to container %s: %s", containerName, strings.Join(cmd, " "))
-		if err := container.Exec(containerName, cmd, nil, &stdout, &stderr); err != nil {
-			return fmt.Errorf("failed to add route '%s' to container %s: %w", strings.Join(cmd, " "), containerName, err)
-		}
-		routesAdded++
-	}
-
-	if routesAdded == 0 {
-		return fmt.Errorf("no valid service CIDRs found to configure routing")
-	}
-
-	return nil
 }
